Debug API replay processing works with Celery tasks
For the debug API, an uploaded file is saved on disk using Python's
tempfile.NamedTemporaryFile with automatic deletion disabled. The
name of this file is passed to a LocallyStoredReplayParsing Celery
task. This task reads the file and returns the data to Celery, which
stores it in Redis.

The response to the user uploading a file is to reply with a URL
to a results page. The GET parameter of the results page is the
celery task ID, which can be used to look up the status and returned
data from the Celery task. When processing completes, the JSON
object on this page turns from status=PENDING to status=SUCCESS,
with the processing results directly in the 'data' property.

The production mode API will work slightly differently. Instead of
storing data locally, the user will upload the file to S3 and
let the server know it has done so. The Celery task launched for
this type will access the file on S3, process it, and write the
resulting JSON file back to S3. The SUCCESS response on the result
page will be the address of the JSON file and not its contents.