Working version of uploading replays and results to S3
This includes gzip compression of the results data. The url that
gets sent back to the requester automatically sets the
content-encoding and content-type headers so that browsers will
automatically uncompress the results. This is so that AJAX calls
can read data directly from S3, but they only have to transact the
compressed data.